+++ XOSL/INCLUDE/Disk.h
@@ -27,14 +27,14 @@ class CDisk {
 
 		int DriveCount(int Fixed);
 
-		int Map(int Drive, unsigned long StartSector);
+		int Map(int Drive, unsigned long long StartSector);
 
-		int Read(unsigned long Sector, void *Buffer, int Count);
-		int Write(unsigned long Sector, void *Buffer, int Count);
-		int Verify(unsigned long Sector, int Count);
+		int Read(unsigned long long Sector, void *Buffer, int Count);
+		int Write(unsigned long  long Sector, void *Buffer, int Count);
+		int Verify(unsigned long  long Sector, int Count);
 
 	private:
-		int Transfer(int Action, unsigned long Sector, void *Buffer, int Count);
+		int Transfer(int Action, unsigned long long Sector, void *Buffer, int Count);
 		void Sector2CHS(unsigned long RSector, unsigned short &SectCyl, unsigned short &DrvHead);
 
 		CDiskAccess DiskAccess;
